refactor : User -> Customer
