+++ voucher-management/src/main/java/com/voucher/vouchermanagement/repository/utils/BlackListDeserializer.java
@@ -1,19 +1,19 @@
 package com.voucher.vouchermanagement.repository.utils;
 
-import com.voucher.vouchermanagement.model.user.User;
+import com.voucher.vouchermanagement.model.customer.Customer;
 import java.util.StringTokenizer;
 import java.util.UUID;
 import org.springframework.stereotype.Component;
 
 @Component("blacklistDeserializer")
-public class BlackListDeserializer implements CsvDeserializer<User> {
+public class BlackListDeserializer implements CsvDeserializer<Customer> {
 
   @Override
-  public User deserialize(String csvLine) {
+  public Customer deserialize(String csvLine) {
     StringTokenizer stringTokenizer = new StringTokenizer(csvLine, ",");
     String id = stringTokenizer.nextToken().trim();
     String name = stringTokenizer.nextToken().trim();
 
-    return new User(UUID.fromString(id), name);
+    return new Customer(UUID.fromString(id), name);
   }
 }
